+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ### Added
 - display the next four clock changes for configured time values with (unpinned) DST in info
 - date and time formats configurable (preset and custom options)
-- `example_configs` and [config manual](https://github.com/RealNoitasinagro/epoch/blob/develop/example_configs/CONFIG_MANUAL.md) on GitHub
+- day quarter indicator in different colors (on/off-switch in Settings)
 - support for direct config loading on Linux commandline
+- `example_configs` and [config manual](https://github.com/RealNoitasinagro/epoch/blob/develop/example_configs/CONFIG_MANUAL.md) on GitHub
 
 ### Changed
 - a few common timezone abbreviations localized for German (MEZ, MESZ, OEZ, OESZ, WEZ, WESZ)